pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codebench"
version = "0.1.0"
description = "Benchmark and evaluation framework for coding LLMs and coding agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0,<3.0",
    "typer>=0.12,<1.0",
    "httpx>=0.27,<1.0",
    "rich>=13.0,<14.0",
    "docker>=7.0,<8.0",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "pytest-cov>=5.0,<6.0",
    "ruff>=0.4,<1.0",
    "mypy>=1.10,<2.0",
    "types-PyYAML>=6.0",
    "types-docker>=7.0",
]

[project.scripts]
codebench = "codebench.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/codebench"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = [
    "B008",    # function-call-in-default-argument — required by typer
    "SIM108",  # if-else-block-instead-of-if-exp — readability preference
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "e2e: end-to-end tests",
]
asyncio_mode = "auto"
